# frontend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mesa de ayuda · IDL",
    "version": "1.0.0",
    "description": "Fechas UTC con Z/offset. Días y presentación America/Montevideo. Bearer JWT y OTP por correo; sin refresh. Estados y transiciones disponibles en catalogs/display/config."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
    "/admin/org-units": {
      "get": {
        "operationId": "orgUnitGet",
        "summary": "Administrar registros",
        "description": "Administrar registros. Códigos inmutables; desactivación lógica. Cambiar rol/actividad invalida sesiones. Se conserva al menos un admin activo. No hay borrado físico.",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "description": "Trazabilidad por petición; no garantiza idempotencia."
          }
        ],
        "responses": {
          "200": {
            "description": "Operación confirmada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrgUnitList"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "request_id": {
                          "type": "string",
                          "readOnly": true
                        }
                      },
                      "required": [
                        "request_id"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "ok",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              }
            }
          },
          "401": {
            "description": "Sesión inválida, expirada o revocada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "403": {
            "description": "Rol sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "404": {
            "description": "Registro o ruta no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "409": {
            "description": "Conflicto de versión, idempotencia o unicidad",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "422": {
            "description": "Validación o transición inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "429": {
            "description": "Cuota de cuenta/IP o cooldown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          }
        }
      },
      "post": {
        "operationId": "orgUnitPost",
        "summary": "Administrar registros",
        "description": "Administrar registros. Códigos inmutables; desactivación lógica. Cambiar rol/actividad invalida sesiones. Se conserva al menos un admin activo. No hay borrado físico.",
        "x-roles": [
          "admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-Request-ID",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 64
            },
            "description": "Trazabilidad por petición; no garantiza idempotencia."
          }
        ],
        "responses": {
          "201": {
            "description": "Operación confirmada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/OrgUnit"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "request_id": {
                          "type": "string",
                          "readOnly": true
                        }
                      },
                      "required": [
                        "request_id"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "ok",
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              }
            }
          },
          "401": {
            "description": "Sesión inválida, expirada o revocada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "403": {
            "description": "Rol sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "404": {
            "description": "Registro o ruta no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "409": {
            "description": "Conflicto de versión, idempotencia o unicidad",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "422": {
            "description": "Validación o transición inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "429": {
            "description": "Cuota de cuenta/IP o cooldown",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          },
          "503": {
            "description": "Dependencia no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string",
                  "readOnly": true
                }
              },
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Segundos de espera para 429"
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitCreate"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrgUnitCreate": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^[A-Z0-9_-]{1,32}$",
            "description": "Código legible; en catálogos es único e inmutable."
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160,
            "description": "Etiqueta descriptiva actual."
          },
          "kind": {
            "type": "string",
            "enum": [
              "area",
              "office",
              "municipality"
            ],
            "description": "Campo del resultado documentado de esta operación."
          },
          "is_active": {
            "type": "boolean",
            "default": true,
            "description": "Desactivación lógica: conserva referencias; usuarios desactivados pierden sesiones."
          },
          "can_receive_tickets": {
            "type": "boolean",
            "default": false,
            "description": "Habilitación explícita como destino, editable por admin."
          },
          "parent_id": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "default": null,
            "description": "Relación opcional de unidad; se rechazan ciclos."
          }
        },
        "required": [
          "code",
          "name",
          "kind"
        ],
        "additionalProperties": false
      },
...
      "OrgUnit": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "readOnly": true,
            "description": "Identificador estable asignado por la base de datos."
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z0-9_-]{1,32}$",
            "description": "Código legible; en catálogos es único e inmutable."
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 160,
            "description": "Etiqueta descriptiva actual."
          },
          "kind": {
            "type": "string",
            "enum": [
              "area",
              "office",
              "municipality"
            ],
            "description": "Campo del resultado documentado de esta operación."
          },
          "is_active": {
            "type": "boolean",
            "default": true,
            "description": "Desactivación lógica: conserva referencias; usuarios desactivados pierden sesiones."
          },
          "can_receive_tickets": {
            "type": "boolean",
            "default": false,
            "description": "Habilitación explícita como destino, editable por admin."
          },
          "parent_id": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "default": null,
            "description": "Relación opcional de unidad; se rechazan ciclos."
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "kind",
          "is_active",
          "can_receive_tickets",
          "parent_id"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "ok": {
            "const": false,
            "description": "Campo del resultado documentado de esta operación."
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "readOnly": true
              },
              "message": {
                "type": "string",
                "readOnly": true
              },
              "errors": {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "readOnly": true
                  }
                }
              },
              "details": {}
            },
            "required": [
              "code",
              "message",
              "errors"
            ],
            "additionalProperties": false,
            "description": "Campo del resultado documentado de esta operación."
          },
          "meta": {
            "type": "object",
            "properties": {
              "request_id": {
                "type": "string",
                "readOnly": true
              }
            },
            "required": [
              "request_id"
            ],
            "additionalProperties": false,
            "description": "Campo del resultado documentado de esta operación."
          }
        },
        "required": [
          "ok",
          "error",
          "meta"
        ],
        "additionalProperties": false
      },
...
      "OrgUnitList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrgUnit"
            },
            "description": "Campo del resultado documentado de esta operación."
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}